// src/types.ts
export type SearchInputType = {
  city: string;
};

export type CityData = {
  country: "GB";
  lat: number;
  local_names: object;
  lon: number;
  name: string;
  state: string;
};

export type WeatherData = {
  cod: string;
  city: {
    id: number;
    name: string;
    coord: {
      lat: number;
      lon: number;
    };
    country: string;
    population: number;
    timezone: number;
    sunrise: number;
    sunset: number;
  };
  list: {
    dt: number;
    main: {
      temp: number;
      feels_like: number;
      temp_min: number;
      temp_max: number;
      pressure: number;
      sea_level: number;
      grnd_level: number;
      humidity: number;
      temp_kf: number;
    };
    weather: {
      id: number;
      main: string;
      description: string;
      icon: string;
    }[];
    clouds: {
      all: number;
    };
    wind: {
      speed: number;
      deg: number;
      gust: number;
    };
    visibility: number;
    pop: number;
    sys: {
      pod: string;
    };
    dt_txt: string;
  }[];
};

export type WeatherDisplayData =
  | {
      date: string;
      minTemperature: number;
      maxTemperature: number;
    }
  | {
      date: string;
      pressure: number;
    }
  | {
      date: string;
      temperature: number;
    }
  | {
      date: string;
      humidity: number;
    }
  | {
      date: string;
      wind: number;
    }
  | undefined;